Add first working CLI11 implementation (Makefile+ simple commander)
